Implement 2D convolution kernel on GPU
In this commit, I have written a kernel to perform 2D convolution by
parallising across multiple threads of the GPU. I have also created a test
case file to validate the 2D convolution kernel. For this I have made a
folder named 'test_images' which contains the input images to test on.
The point spread function is user defined and can be changed from the
test case script. The time taken to perform the 2d convolution using the
inbuilt scipy.signal.convolve2d function on the CPU is ~20 ms while that
performed using the nuba kernel is about 4ms. About 1/4 the time taken
by the cpu inbuilt function. Right now, the test case has been validated
for a 2d matrix image but when a 2d matrix image (RGB) is given as a
test case, the output doesn't seem to be correct. Need to debug this.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>