Fixed bug in 2d convolution algo on GPU
In this commit, I have fixed a major bug in the way the 2d convolution
was being computed on the GPU using the numba cuda kernel. The following
are the changes made which fixed the bug.
1. BlocksperGrid varaible was not being computed correctly. The ceil
operation should be applied after division with the threadsPerBlock but
it was applied before the division.

2. The point spread function (2d impulse response) was not being
accessed correctly and hence it had some zero values. Because of this,
the convolution sum was not correct.

3. The output image was not indexed correctly with respect to the
thread indices

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>