Compute 2D convolution for any 3D image
In this commit, I have solved the previous issue where I was able to
perform a 2D convolution only for a 2D image. Now, with these changes,
the convolution can be generalized to any 3D image which has the 3
components of Red, Green, Blue (RGB). Following are the changes
introduced.

1. Performing 2D convolution to each of the 2D slices of the 3D image
i.e. 2D convolution with each of the R image, G image, B image
separately.

2. Previously, I was not able to generalise to a 3D image. Now I have
solved it. This is the explanation.
Generally, the input images are uint8 i.e each pixel value in the image
is represented by a unsigned 8 bit integer value and hence it will have
integer values from 0 to 2^5 -1 i.e. 255. But, our convolution operation
might result in values which are larger than 255. Hence after our
convolution operation, we need to convert the image from float32 to
uint8. For this, we first normalize, the every slice of the output 3D
image with the corresponding maximum value. Now each slice will have
values from 0 to 1. Now, we multiply with 255 and hence each slice has
values from 0 to 255. After this, we convert to uint8. Now we can plot
and compare with respect to the nput image.

3. Also added the time taken by the CPU and GPU in the results plot.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>