Implement edge detection on GPU
In this commit, I have successfully implemented the edge detection
algorithm on both the CPU and GPU. The outputs from CPU and GPU were
exactly matching. This is expected because, I had already bitmatched the
2D convolution kernel with the signal.convlve2d library function which
runs on CPU. Since this edge detection also essentially involves a 2-D
convoltion, it was anyways expected that the GPU version and CPU version
would match. Also, the time taken by the GPU kernel was about 5 ms where
as the time taken by the CPU version was 40 ms. The GPU version seems to
be faster than the CPU version by a factor of close to 8. The edge
detection is done by taking a horizontal gradient and a vertical
gradient and then taking the inner product with a cos and sin function
respectively for different sets of angles and then taking the max across
different angle hypothesis. The edge detection can also be done by
taking the l-2 norm of the horizontal gradient image and vertical
gradient image. Both these methods seems to be giving similar results.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>