Compare 1D CFAR CA on CPU and GPU
In this commit, I have coded the 1D CFAR CA algorithm on the GPU and
verified against the CFAR CPU libraries that I had already built earlier in
the Spectral Estimation repository. The time taken by the CFAR CA on the
CPU for a 1024 length signal is about 11 ms while the time taken by the
GPU kernel is only 1ms. In the next commit, I will also code the 1D CFAR
OS on GPU.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>