Fix major GPU coding bug for CFAR CA
In this commit, I have fixed a major bug with respect to GPU coding.
Whenever, you need a scratchpad variable(matrix) for intermediate computes on
the GPU kernal, always pass it as a global memory access such that each thread
has access to its own row/column of the scratchpad array. This is not
required if the intermediate variable is a 1x1 variable. If you pass a
common scratchpad memory to all the threads, there might be conflicts
since all threads have to access the memory simultaneously and this
might result in faulty results at times.
Also, preferably, allot each row of a 2D scratchpad variable to individual thre d instead of columns, since this will allow contiguous memory access for each
thread.

The following are the changes made to the scripts after taking above
into consideration.
1. In the CFAR_lib file, I have made sure that the datatypes of the
variables for the CPU and the GPU version are the same so that the CPU
and GPU outputs match exactly.
2. In the numbaCudaKernels.py file, I'm now passing the scratchpad
variable as a 2D array.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>