Compare timings of CFAR OS on CPU and GPU
In this commit, I have brought up the numba cuda kernel for the CFAR OS
as well. I have also matched it with the CPU version. The
'cfar_1dTestcase.py' script compares the CPU and GPU timing performance
of CFAR CA and CFAR OS. I have also ensured that the the CPU and the GPU
versions of the CFAR OS and CA are bit matched.
Also, I have changed the size of the scratchpad memory to num_threads x
num_noise_samples, where num_threads = length of the original signal. In
my earlier commit, the size of the scratchpad array was
length_extended_signal x num_noise_samples. This was unnecessary because
we need to allocate a scratch pad memory for only len(original_signal) #
of threads.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>