Perform 2D CFAR CA cross on GPU
In this commit, I have written a numba cuda kernel to perform 2D CFAR CA cross on GPU. I have also added a test script named 'cfar_2dTestcase' to validate
the performance and timings of the 2D CFAR CA cross implemented on CPU
and GPU. The CFAR CA on the CPU takes close to 3.5 seconds or 3500 ms
whereas the time taken by the cuda kernel of the CFAR CA is about 8 ms!
We observe a speed improvement of the order of 400 i.e. the GPU version
of 2D CFAR is around 400 times faster than the CPU version.

Also, Another point I've observed with the numba cuda kernels on the GPU
is that they are very sensitive to datatype precision values. For
example float 32 number of the order of 1e-8 on the CPU when pushed to
GPU is quantized to 0. In my test case examples for the image on which
CFAR has to run, the noise values very were tiny because of very low
noise power(of the order of 1e-8) and hence when these values when
inside the GPU kernel were getting quantized to 0. So in order to avoid
this, I have raised the noise floor values and the signal values by 30
dB. Now the results of the CPU and GPU seem to match.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>