Compute 1D convoultion using C based pyCUDA kernel
In this commit, I have written my first C-based pyCUDA kernel. I have
implemented the 1D convulution in pyCUDA and bitmatched it with the
numba CUDA version. Also, I have created a separated file for pyCUDA
kernels called 'pyCudaKernels.py'. I have also made separate
repositories for the pyCUDA test cases and numba CUDA test cases. The
same test cases cannot be used for the both the pyCUDA implementation as
well as the numba CUDA implmentations. This is because of the conflict
between the cupy libraries and the pyCUDA libraries. The pyCUDA kernels
can be run immediately after the numba CUDA kernels are run but the
ohter way around is not possible i.e. if you try to run a numba CUDA
kernel after a pyCUDA kernel, it throws up CUDA context issues which
cannnot be operated upon by numba. Hence the need for separate test
cases for each of them.

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>