Softmax implementation for 10 million elements
In this commit, I have implemented 2 flavours of softmax function specifically
catering to very large number of elements ~ 10 million.
Flavour 1: When the compute memory is constrained

When we cannot accomodate 10 million elements in one go to compute the softmax, we perform
a tiling or sliding window based method of computing the maximum value per tile/window and
then use this to compute the local denominator sum. We then move to the next window and update the
max value and the denominator sum. This approach is similar to the Flash Attention for 1D case.

Flavour 2: When compute units are constrained

Softmax requires normalizing each value by the largest value to ensure stability of the exponential function.
Computing max element in an array is O(N) and when N is large (like 10 million), this can be huge.
This is where parallelization on GPU helps. By launching k threads, we can cut down this compute to
O(N/k). Each thread works on a chunk of data and computes the local maximum and the local denominator sum
using the local maxima. Then the outputs from each of the threads can be combined either again on GPU or
even on CPU (since by now majority of the compute is already done) to obatain the global maxima and denominator sum

I have implemented both these flavours of softmax and compared against the pytorch implementation.
I have used different metrics to compare the performance like:
    1. Average error per sample
    2. Sum of softmax must sum to 1
    3. KL divergence between true pytorch implementation and each method

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>