Implement GPU kernel using shared memory
In this commit, I have implemented softmax on the GPU using the concept of shared memory!
This is the first time Im using shared memory which is shared across threads of a block.
This code has softmax implementations using global memory as well as shared memory.
However, when I run this code on my local machine, both the global memory based implementation as well as
shared memory implementation are taking a lot more time(>200ms)! But when I run the same code on google colab,
on a A100 GPU, Im able to hit sub 2ms with the shared memory implementation and sub5ms with the
global memory implementation! This could be because, on the local machine,
there could be several processes parallely running slowing down the speed. But the global memory implementation
should be taking same time as previously, but in this script, that too has slowed down.
The same implementation in the other script 'softmax_gpu_globalmemory' is running quite fast!
Need to debug this issue.

The google colab link where Im able to hit 2ms for shared memory implementation and 4ms for
global memory implementation is available here:
    https://colab.research.google.com/drive/1i_pTjXEjyDw-bBYq0xwV6fs1MtFz_vC3#scrollTo=DkzAMq1BqJSs
"""

Signed-off-by: Sai Gunaranjan Pelluri <saigunaranjanpelluri@gmail.com>